HW Project - 2018 Spring
The idea was to create a model to predict the product based on the consumer complaint narrative. For this I used unstructured data analytics techniques- the textual consumer complaints were first parsed, tokenized , lemmatized. Then a TFID matrix was created. Following this a model was created. to predict the product / sub-product for which the complaint can be attributed to. i used a tuned random forest model for this.